+++ examples/counter/src/main.rs
@@ -4,16 +4,15 @@ use axum::{Router, response::IntoResponse};
 use vixen::{
     action, fragment,
     maud::{DOCTYPE, Markup, html},
-    partial,
+    partial, route,
     routing::RouterExt,
-    view_path,
 };
 
 const ADDR: &str = "127.0.0.1:4002";
 
 static COUNT: AtomicI64 = AtomicI64::new(0);
 
-#[view_path("/")]
+#[route("/")]
 struct HomePath;
 
 async fn home(_: HomePath) -> Markup {